lin/requirements.txt
PyGObject>=3.42
PyOpenGL>=3.1.6
Pillow>=9.0
numpy>=1.21
